Refactored the code to seperate out context and its method from Tool.jsx && Deletion of all APIs, certs is now possible
