+++ src/components/EnvironmentVarView.jsx
@@ -4,7 +4,7 @@
  */
 
 import React, { useContext, useEffect, useRef, useState } from "react";
-import { GlobalStore } from "../contexts/GlobalContext";
+import GlobalContextProvider, { GlobalStore } from "../contexts/GlobalContext";
 import { Button, TextInput } from "@carbon/react";
 import { Add, ArrowRight, Subtract } from "@carbon/react/icons";
